style(sql_app): divided sql_app_main into model and pd_db
